Add users endpoint
