+++ apps/api/src/modules/example/dashboard.ts
@@ -1,6 +1,7 @@
 import { z } from "@yas/validate";
 import { t } from "../../definition/trpc";
 import { dashboardSchema } from "./types";
+import { createRandomUsers, createSeededRandom } from "./fixtures";
 
 export function createDashboardProcedure() {
   return t.procedure
@@ -32,52 +33,7 @@ export function createDashboardProcedure() {
           { name: "Nov", value: rand() * 5600 },
           { name: "Dec", value: rand() * 1337 },
         ],
-        recentSales: [
-          {
-            name: "Olivia Martin",
-            email: "olivia.martin@email.com",
-            amount: rand() * 1999,
-            avatarUrl: randomAvatarUrl(0),
-          },
-          {
-            name: "Jackson Lee",
-            email: "jackson.lee@email.com",
-            amount: rand() * 39,
-            avatarUrl: randomAvatarUrl(1),
-          },
-          {
-            name: "Isabella Nguyen",
-            email: "isabella.nguyen@email.com",
-            amount: rand() * 299,
-            avatarUrl: randomAvatarUrl(2),
-          },
-          {
-            name: "William Kim",
-            email: "will@email.com",
-            amount: rand() * 99,
-            avatarUrl: randomAvatarUrl(3),
-          },
-          {
-            name: "Sofia Davis",
-            email: "sofia.davis@email.com",
-            amount: rand() * 39,
-            avatarUrl: randomAvatarUrl(4),
-          },
-        ],
+        recentSales: createRandomUsers(rand),
       };
     });
-}
-
-const randomAvatarUrl = (index: number) =>
-  `https://picsum.photos/40/40?grayscale&random=${index}`;
-
-function createSeededRandom(seedStr: string) {
-  let seed = Array.from(seedStr).reduce(
-    (sum, char) => sum + char.charCodeAt(0),
-    0,
-  );
-  return () => {
-    const x = Math.sin(seed++) * 10000;
-    return x - Math.floor(x);
-  };
 }
