+++ apps/api/src/modules/example/router.ts
@@ -1,9 +1,11 @@
 import { z } from "@yas/validate";
 import { t } from "../../definition/trpc";
 import { createDashboardProcedure } from "./dashboard";
+import { createUsersProcedure } from "./users";
 
 export function createExampleRouter() {
   return t.router({
+    users: createUsersProcedure(),
     dashboard: createDashboardProcedure(),
     hello: t.procedure
       .input(z.string())
